settings.py: change default value for `DJANGO_DEBUG` to `True` [settings][dx]
Change the default value of `DEBUG` to `True`.

Since I plan to release bugpile as an Open Source project, I should
make default settings correspond to a local development setup, as
suggested in [1]:

   "If you’re releasing the source code for your project, a common
   practice is to publish suitable settings for development, and to
   use a private settings module for production."

This will make it as easy as possible for other people to clone the
project and try it out locally.

[1]: https://docs.djangoproject.com/en/5.2/howto/deployment/checklist/