import: `legal/terms_of_service.md` [wip][legal]
My initial (incomplete) work on a Terms of Service document.

I started with a ChatGPT-generated outline of the GoFundMe Terms of
Service [1], and now I'm part way through the process of
modifying/adapting it for BugPile. As of this commit, I've reached the
start of the "Prohibited User Conduct" section

Even though, the document is incomplete, I'm starting to get antsy
about putting it under version control and backing it up, hence this
commit.

[1]: https://chatgpt.com/share/686d336d-6dd0-8006-b6d1-40476d507aa3