paypal: fix blank pop-up window [bugfix]
Set `SECURE_CROSS_ORIGIN_OPENER_POLICY = 'same-origin-allow-popups'`.

This is a security compromise which increases vulnerability to
"XS-Leak" attacks, but it's not that big of a risk, and it seems like
the only way to get the PayPal payment button/pop-up to work properly.

See the following issue for details/discussion:
https://github.com/paypal/paypal-js/issues/534#issuecomment-2230393690