apps.py: only call `init_paypal` when running `./manage.py runserver` [bugfix]
`init_paypal` was being called on every management command (e.g.
`./manage.py makemigrations`, `./manage.py migrate`), whereas it
should only be called when starting the server (i.e. `./manage.py
runserver`).

It's not a big deal, but it was causing unnecessary calls to the
PayPal API, and cluttering the Django log with multiple "PayPal
configured for sandbox mode" messages.