paypal.py: server-side implementation of "capture" step [feat]
Move the implementation of the final "capture" step of a PayPal payment from
client-side Javascript to a server-side
webhook (`/paypal/capture_order`).

The "capture" step is the final approval by the seller, after the user
has authorized the payment on the PayPal website. The reason we want
to handle this step on the server-side is not for better security, but
because it provides us with a hook for updating donation total of the
target GitHub issue in the database. (I have not implemented the
actual updating of the donation total yet, though.)