repo_issues.html: remove search box [simplify]
In the spirit of minimalism and simplicity, remove the search box for
now. Up until now, the search box was just a non-functional
placeholder anyway.

At least for my own usage, I don't think search filtering or
pagination is really needed. It's fine to just list all of the issues
on one page, and if users want to search for specific issues, they can
use Ctrl+F.