base.html: implement user dropdown menu (like GitHub) [feat][UX]
Instead of showing the signed-in user's avatar icon, GitHub username,
and "Logout" button directly in the nav bar, just show an avatar icon
with a dropdown menu (like GitHub does).

With the previous layout, there were too many items on the nav bar and
it would wrap to the next line on mobile, in an ugly way.  In
addition, the new dropdown menu feels more conventional/familiar than
the previous layout.