index.html: remove outdated sections (How it Works, Waitlist) [doc][cleanup]
Remove outdated "How it Works" and "Waitlist" sections.

The "How it Works" explanation is outdated, since it was written when
I was planning to take direct Stripe/PayPal payments on the website,
rather than integrating with GitHub Sponsors. In addition, I plan to
put "How it Works"-type information on the FAQ page (yet to be
created).

And I no longer plan to have a waitlist, so the "Waitlist" section
doesn't apply anymore.