repo_issues: change "Add or Remove Funds" button to outline [style]
Change the "Add or Remove Funds" button to an outline style by
default, but change to a filled style on hover, like the GitHub
Sponsors button in the header.

Before, the "Add or Remove Funds" looked like it had too much weight
to it, in comparison to the other page elements.