repo_issues: use issue label colors from GitHub issue data [UX]
Use the same issue label colors that GitHub uses. This is not hard to
do, because the label colors part of the JSON data for each GitHub
issue in the database. But the initial LLM-generated code used some
weird custom logic based on whether the label text contained "bug",
"feature", "priority", etc.

One tricky thing is that it's not totally straightforward to pick a
text color that is readable on top of the various issue label colors
-- white text works well on dark colors and black text works well on
light colors. The LLM did something pretty clever here, where it
calculated the luminance of the label color and then set the
foreground text to black or white based on the luminance value. Pretty
nice.