repo_issue: change buttons to dark text on light background [style]
I felt that the old style of the "Add or Remove Funds" button (large
dark blue button with white text) was too intense/eye-catching.

Also, this commit makes the styling of the GitHub Sponsors button and
the "Add or Remove Funds" buttons consistent, which makes the web page
look more harmonious overall. Previously, the GitHub Sponsors button
was an outline style, with dark pink text and border on white
background, and the "Add or Remove Funds button" was an inverted text
style, with white text on a dark blue background.