index.html: show inline screenshot on mobile [style]
On mobile, show the screenshot inline with the text, after the first
paragraph. But on desktop (screen width > 768px), keep the current
layout with the screenshot in its own column to the right of the text.

Previously, the layout was bugged on mobile. The screenshot would
appear after the end of the intro text (too late, IMO), and there was
a large (ugly) vertical gap between the text and the image.

Implementation note: I don't think its possible to implement different
layout positions for the same <img> tag based on screen width, so this
commit uses a clever trick -- it adds two different <img> tags, and
shows one or the other based on the current screen width (i.e. a CSS
media query).