repo_issues: highlight target issue with gray background [style]
When following a link to a specific issue on the repo issues page,
e.g. from the "Trending Issues" list on the homepage, highlight the
target issue with a light gray background.

Linking to a specific issue on the repo issues page is done with an
HTML anchor (e.g.
`http://localhost:8000/gh/benvvalk/qutebrowser/issues#issue-16`).