settings.py: implement `ALLOWED_GITHUB_USERS` setting [setting]
Add a `ALLOWED_GITHUB_USERS` setting that restricts logins to a
specific list of GitHub usernames (e.g. the developer's own GitHub
account(s)). This is useful if locking things down while the site is
under development or maintenance.

If `ALLOWED_GITHUB_USERS` is not set, allow anyone to sign in.