repo_issues: various fixes when signed-in user == repo owner [bugfix]
If the signed-in user is the github repo owner:

* Disable the "Add or Remove Funds" button
* Hide the "You:" amount under the donation total (since it will
always be zero)
* Don't show the "you haven't donated yet" reminder at the top of the
issues list
* On the server side, forbid self-donations

I don't know if it's actually possible for a user to donate to
themselves on GitHub Sponsors, and I don't want to risk getting my
GitHub account shut down to find out.

There is not much incentive for a user to sponsor their own issues,
although I suppose they might to do it to create a false impression of
popularity, in the hopes of encouraging real donations.

There is not much I can do to prevent such shenanigans because the
user that wants to donate to themselves can just create a second
GitHub account and donate from there instead. (In fact, that is
exactly what I had to do to test the website.)