owner_issues: fix error when clicking "Add or Remove Funds" [bugfix]
Fix `Uncaught ReferenceError: False is not defined` when clicking "Add
or Remove Funds" button.

The problem was that I was passing the Python `issue` verbatim to the
`openModal` Javascript function in the following line:

`<button class="donate-btn" onclick="openModal({{ issue }}) ...`

That doesn't work because Javascript considers `False` to be an
undefined variable in `is_selected: False`.

This commit refactors the HTML to pass the `openModal` arguments using
`data-` attributes on the `<button>` tag instead, which is much
cleaner IMO.