sync_github_issues: add `--loop` for continuous syncing [feat]
Add a `--loop` option so that I easily define a DigitalOcean "worker"
job that continuously syncs issue data with GitHub.  Note that this
job is just a fallback for when we miss webhooks events, so the sync
doesn't necessarily have to happen quickly or frequently.

I also added `--loop-delay=SECONDS` option for specifying a delay
between sync cycles (with a default value of 0). Note that using the
default `loop-delay` value of 0 will not result in API abuse, because
there are still delays between the individual API requests within each
sync cycle. (The inter-request delays are currently hard-coded to use
a random value in the range of 2-10 seconds.)