sync: don't delete issues with existing funding [bugfix][sync]
Previously the webhook was unconditionally deleting issues from the
database if either:

(1) The issue was closed, *OR*
(2) The `sponsoredissues.org` label was removed

However, we only want to delete the issue from the database if it does
not yet have any funding from users. In the case that the issue *does*
already have funding, we need to keep the issue in our database for
multiple reasons:

(1) It allows to compute interesting historical stats, such as
average funding amount for resolved issues, average time to resolve
issues, etc.

(2) If the maintainer reopens a closed issue, we need to restore
the funding amount to the same value as when it was closed.

(3) If the maintainer accidentally removes the `sponsoredissues.org`
label from an issue with non-zero funding, we don't want to
automatically return all the assigned funds to users. Instead,
we continue to show the issue on the maintainer's sponsored issues
page, but in a special frozen/error state with the "Add or Remove
Funds" button disabled and an explanatory error message.