sync_github_issues: update all funded issues, regardless of labels [bugfix][sync]
Modify `sync_github_issues.py` to retrieve the latest data for all
funded issues, including issues that no longer have the
`sponsoredissues.org` issue label on GitHub.

This is necessary to gracefully handle the edge case where the
maintainer accidentally removes the `sponsoredissues.org` label from
an issue that already has funding from users. In such a case, we don't
want to just delete the issue from the sponsoredissues.org database,
all the assigned funds to be returned to contributors without any
warning or explanation.

Instead, I plan to display the issue in a special "frozen" state, with
an explanatory message and the "Add or Remove Funds" button
disabled. I haven't implemented the special display rules/styling for
such issues, but I plan to do so soon.