sync_github_issues: remove `--limit` (`repo_limit` parameter) [simplify]
Simplify the code by removing the `--limit` option and the associated
`repo_limit` parameter to various methods.

The doc string for the `--limit` option was wrong. It doesn't limit
the total number of repos that are queried per app installation, but
rather how many repos are batched into a single GraphQL
query/response. Regardless of how the queries are batched,
`github_sync_issues` always queries/updates *all* of the repos for an
app installation.

Rather than having `--limit` option to control batching, I just set it
to a hard-coded value of 30 repos-per-query for now.